Data engine v2 (#1817)
