+++ nautilus_core/common/src/lib.rs
@@ -47,5 +47,6 @@ pub mod xrate;
 #[cfg(feature = "ffi")]
 pub mod ffi;
 
+pub mod client;
 #[cfg(feature = "python")]
 pub mod python;
