+++ nautilus_core/common/src/messages/data.rs
@@ -17,12 +17,12 @@ use std::{any::Any, sync::Arc};
 
 use nautilus_core::{nanos::UnixNanos, uuid::UUID4};
 use nautilus_model::{
-    data::DataType,
+    data::{Data, DataType},
     identifiers::{ClientId, Venue},
 };
 
+// TOOD: redesign data messages for a tighter model
 pub struct DataRequest {
-    pub request_id: UUID4,
     pub correlation_id: UUID4,
     pub client_id: ClientId,
     pub venue: Venue,
@@ -30,19 +30,19 @@ pub struct DataRequest {
     pub ts_init: UnixNanos,
 }
 
+pub type Payload = Arc<dyn Any + Send + Sync>;
+
 pub struct DataResponse {
-    pub response_id: UUID4,
     pub correlation_id: UUID4,
     pub client_id: ClientId,
     pub venue: Venue,
     pub data_type: DataType,
-    pub data: Arc<dyn Any + Send + Sync>,
     pub ts_init: UnixNanos,
+    pub data: Payload,
 }
 
 impl DataResponse {
     pub fn new<T: Any + Send + Sync>(
-        response_id: UUID4,
         correlation_id: UUID4,
         client_id: ClientId,
         venue: Venue,
@@ -51,7 +51,6 @@ impl DataResponse {
         ts_init: UnixNanos,
     ) -> Self {
         Self {
-            response_id,
             correlation_id,
             client_id,
             venue,
@@ -62,16 +61,29 @@ impl DataResponse {
     }
 }
 
-pub enum DataCommandAction {
+#[derive(Debug, Clone, Copy)]
+pub enum Action {
     Subscribe,
-    Unsubscibe,
+    Unsubscribe,
 }
 
-pub struct DataCommand {
+pub struct SubscriptionCommand {
     pub client_id: ClientId,
     pub venue: Venue,
     pub data_type: DataType,
-    pub action: DataCommandAction,
+    pub action: Action,
     pub command_id: UUID4,
     pub ts_init: UnixNanos,
+}
+
+pub enum DataEngineRequest {
+    DataRequest(DataRequest),
+    SubscriptionCommand(SubscriptionCommand),
+}
+
+// TODO: Refine this to reduce disparity between enum sizes
+#[allow(clippy::large_enum_variant)]
+pub enum DataClientResponse {
+    DataResponse(DataResponse),
+    Data(Data),
 }
